Add include directory -option
This allows the user to define one or more include directories from the
command line. If a node does not come from a file that is under one of
those directories, it will not be analysed.

The rationale for the feature is that clang creates AST for all of the
included files, including 3rd party headers. There is no reason to
analyse nodes from them, as they are not necessarily implemented
following the same naming style.