Fix error description for "negative failures"
By "negative failures" I mean failures that fail, because there is a
redundant post- or prefix. E.g. a-prefix, should not be used for
variables that are not arrays.

After this change, the error description is 'non-array variable
"aVariable" has redundant prefix "a"'. The addition was the prefix "non"
in the variable type.