Give error for redundant If-postfix for classes
Note that this reverts part of an earlier commit that tried to simplify
the postfix rules. However, also the more simple can have conditions, so
that a) missing postfix is violation if a condition is true and be)
existing postfix is violation, if condition is false.

An example is If-postfix which needs to be added for interface classes,
but is a violation for other classes.

The current solution duplicates some of the logic from AffixedNameRule.
Hopefully with more experience a more general solution will emerge so we
can remove this duplication.