Automatically pass path to clang builtin headers
According to clang documentation [1], there are some builtin header
files that are expected to be located in a specific folder that is
relative to the used tool.

This works for clang, because clang installation will create the
required directory / header structure. It does not work so well with the
tools that are using clang, but are based in another directory
structure.

The current solution for check_name is to automatically add the include
directory based on the llvm_path option.

[1] http://clang.llvm.org/docs/LibTooling.html#libtooling-builtin-includes