Add input to feature test error output
For example, if test is testing that pointer variable should have
p-prefix, but fails, then the error output will contain also the input
that was given to check_name.

The goal is to make easier to pinpoint the cause of the error. Sometimes
the cause that the generated input is incorrect.